Use runtime classpath at root to workaround Dagger/Hilt API vs Impl issue.
This CL adds a new Hilt option called 'enableClasspathAggregation' that will configure the compile classpath of the app Gradle project to use the runtime classpath of the project. This means that transitive dependencies will be available during compilation which in turn will allow Dagger to traverse the classes along the dependency tree and will allow for Hilt's aggregating classes to be discovered.

The classpath configuration is done by resolving the runtime configuration with an artifact view that simply copies the output jars. This solution is inefficient and will cause build performance impact, but it is a starting point that can be further optimized by using a smarter transform that can extract the necessary classes required by Dagger and Hilt.

Fixes: https://github.com/google/dagger/issues/1991
RELNOTES=Use runtime classpath at root to workaround Dagger/Hilt API vs Impl issue.
PiperOrigin-RevId: 337516449